Start showing multiple columns at 3 columns, instead of 2 (#17)
With apologies to information-density geeks, the page just looks
better with a single column.
The multi-column display can still useful for a big, static (not
scrollable) screen, for example at a conference, but then there
should be 3 or 4 columns. If there's space only for two,
let's not overload the reader, and show the points one by one.